update
git-svn-id: http://svn.tudos.org/repos/oc/tudos/trunk@30 d050ee49-bd90-4346-b210-929a50b99cfc

+++ l4/pkg/bootstrap/server/src/gunzip.c
@@ -166,7 +166,7 @@ static void initialize_tables (void);
 
 static void show_progress(int done, int len)
 {
-  int r = printf("%d%%", (done * 100) / len);
+  int r = printf("%lld%%", ((unsigned long long)done * 100) / len);
   while (r-- > 0)
     putchar('\b');
   fflush(NULL);
